// projects/igniteui-angular/animations/src/interface.ts
export interface IAnimationParams {
    delay: string;
    duration: string;
    easing: any;
    startOpacity?: number;
    endOpacity?: number;
    startAngle?: number;
    endAngle?: number;
    startDistance?: string;
    endDistance?: string;
    fromPosition?: string;
    toPosition?: string;
    fromScale?: number;
    midScale?: number;
    toScale?: number;
    xPos?: string;
    yPos?: string;
    direction?: string;
    rotateX?: number;
    rotateY?: number;
    rotateZ?: number;
    startHeight?: string;
    endHeight?: string;
    startPadding?: string;
    endPadding?: string;
}